blog_run版本11，完善访问权限功能，正搭建忘记密码邮箱验证/requirements.txt
django-haystack
amqp==1.4.9
anyjson==0.3.3
Babel==2.7.0
backports.csv==1.0.7
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11.21
django-celery==3.2.2
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-filter==2.2.0
django-formtools==2.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-redis==4.10.0
django-reversion==3.0.4
djangorestframework==3.10.2
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.25.0
flower==0.9.3
future==0.17.1
greenlet==0.4.15
httplib2==0.9.2
idna==2.8
itsdangerous==1.1.0
jdcal==1.4.1
kombu==3.0.37
Markdown==3.1.1
meld3==1.0.2
monotonic==1.5
mysqlclient==1.4.2.post1
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.1.0
PyMySQL==0.9.3
python-memcached==1.59
pytz==2019.1
PyYAML==5.1.1
redis==2.10.6
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
supervisor==4.0.4
tablib==0.13.0
tornado==5.1.1
urllib3==1.25.3
xadmin==0.6.1
xlrd==1.2.0
xlwt==1.3.0
